requirements.txt
python-dotenv>=1.2.1
openai-agents>=0.6.4
google-auth>=2.36.0
google-auth-oauthlib>=1.2.1
google-auth-httplib2>=0.2.0
google-api-python-client>=2.154.0
tzlocal>=5.0.0